feat: mover schema de post para um novo arquivo e atualizar importações
