+++ components/posts/_components/post-form.tsx
@@ -10,7 +10,7 @@ import { Button } from "@/components/ui/button"
 import { BasicInfoStep } from "./steps/basic-info-step"
 import { ContentStep } from "./steps/content-step"
 import { CategorizationStep } from "../_components/steps/categorization-step"
-import { postSchema, type PostFormData } from "@/schemas/auth"
+import { postSchema, type PostFormData } from "@/schemas/post"
 import type { CreatePostDTO } from "../../../types/post"
 import { createPost } from "@/services/post/post"
 import { toast } from "sonner"
