+++ schemas/auth/index.ts
@@ -3,7 +3,6 @@ import { z } from "zod";
 export const CredentialsSchema = z.object({
 	email: z.string().email(),
 	password: z.string().min(6),
-	code: z.optional(z.string()),
 });
 
 export const RegisterSchema = z.object({
@@ -28,7 +27,6 @@ export const UserSettingsSchema = z
 		email: z.optional(z.string().email()),
 		password: z.optional(z.string().min(6)),
 		newPassword: z.optional(z.string().min(6)),
-		isTwoFactorAuthEnabled: z.optional(z.boolean()),
 	})
 	.refine(
 		(values) => {
@@ -71,16 +69,5 @@ export const NewPasswordSchema = z.object({
 export const MagicLinkSignInSchema = z.object({
 	email: z.string().email(),
 });
-
-
-export const postSchema = z.object({
-	title: z.string().min(3, "Título deve ter pelo menos 3 caracteres"),
-	description: z.string().min(10, "Descrição deve ter pelo menos 10 caracteres"),
-	content: z.string().min(50, "Conteúdo deve ter pelo menos 50 caracteres"),
-	category: z.string().min(1, "Selecione uma categoria"),
-	tags: z.array(z.string()).min(1, "Adicione pelo menos uma tag"),
-})
-
-export type PostFormData = z.infer<typeof postSchema>
