+++ schemas/post/index.ts
@@ -1,0 +1,13 @@
+import { z } from "zod";
+
+export const postSchema = z.object({
+    title: z.string().min(3, "Título deve ter pelo menos 3 caracteres"),
+    description: z.string().min(10, "Descrição deve ter pelo menos 10 caracteres"),
+    content: z.string().min(50, "Conteúdo deve ter pelo menos 50 caracteres"),
+    category: z.string().min(1, "Selecione uma categoria"),
+    tags: z.array(z.string()).min(1, "Adicione pelo menos uma tag"),
+})
+
+export type PostFormData = z.infer<typeof postSchema>
+
+
